requirements.txt
Flask
Flask-WTF
Flask-Bcrypt
Flask-SQLAlchemy
python-dotenv
WTForms
pymysql
cryptography